Currentstate: fix cast on sonar range

+++ CurrentState.cs
@@ -452,7 +452,7 @@ namespace MissionPlanner
         public float accel_cal_z { get; set; }
 
         [DisplayText("Sonar Range (meters)")]
-        public float sonarrange { get { return toDistDisplayUnit(_sonarrange); } set { _sonarrange = value; } }
+        public float sonarrange { get { return (float)toDistDisplayUnit(_sonarrange); } set { _sonarrange = value; } }
         float _sonarrange = 0;
         [DisplayText("Sonar Voltage (Volt)")]
         public float sonarvoltage { get; set; }
